Virtual-Lab-Assistant/requirements.txt
rasa==3.6.21
rasa-sdk==3.6.2
fastapi==0.115.9
uvicorn==0.34.0
mysql-connector-python==9.2.0
python-dotenv==1.0.1
pydantic<=1.10.10
aiohttp==3.9.5
rapidfuzz==3.12.2
requests==2.32.3
openai==1.51.0  
pymupdf==1.25.4
faiss-cpu==1.7.4
httpx==0.27.0
sentence-transformers==3.1.1
python-jose==3.3.0           # 🔐 JWT encoding/decoding
passlib[bcrypt]==1.7.4       